Pulling out date parsing into its own class, adding more checks on final date
