+++ config/boot.rb
@@ -42,6 +42,7 @@ require 'folio/holdings'
 require 'folio/marc_record_instance_mapper'
 require 'folio/marc_record_mapper'
 require 'folio/mhld_builder'
+require 'folio/sortable_date_parser'
 require 'folio/status_current_location'
 require 'folio/types'
 require 'libraries_map'
