+++ spec/lib/folio/holdings_spec.rb
@@ -44,7 +44,7 @@ RSpec.describe Folio::Holdings do
       end
     end
 
-    context 'when cronology is an empty string' do
+    context 'when chronology is an empty string' do
       let(:target) { { 'chronology' => '', 'enumeration' => 'Episode 8' } }
       let(:options) { [{ 'chronology' => '', 'enumeration' => 'Episode 7' }, target, { 'chronology' => '', 'enumeration' => 'Episode 2' }] }
 
@@ -53,36 +53,9 @@ RSpec.describe Folio::Holdings do
       end
     end
 
-    context 'when comparing cronology with a date to one that is an empty string' do
+    context 'when comparing chronology with a date to one that is an empty string' do
       let(:target) { { 'chronology' => '2003', 'enumeration' => '' } }
       let(:options) { [{ 'chronology' => '', 'enumeration' => '' }, target, { 'chronology' => '', 'enumeration' => '' }] }
-
-      it 'selects the target' do
-        expect(latest).to include(target)
-      end
-    end
-
-    context 'when enumeration has multiple integers' do
-      let(:target) { { 'chronology' => '', 'enumeration' => 'bd.54 t.2' } }
-      let(:options) { [{ 'chronology' => '', 'enumeration' => 'bd.54 t.1' }, target, { 'chronology' => '', 'enumeration' => 'bd.53 t.1' }] }
-
-      it 'selects the target' do
-        expect(latest).to include(target)
-      end
-    end
-
-    context 'when enumeration is text' do
-      let(:target) { { 'chronology' => '', 'enumeration' => 'ccc' } }
-      let(:options) { [{ 'chronology' => '', 'enumeration' => 'bbb' }, target, { 'chronology' => '', 'enumeration' => 'aaa' }] }
-
-      it 'selects the target' do
-        expect(latest).to include(target)
-      end
-    end
-
-    context 'when enumeration is mixed' do
-      let(:target) { { 'chronology' => '', 'enumeration' => '52' } }
-      let(:options) { [{ 'chronology' => '', 'enumeration' => 'bbb' }, target, { 'chronology' => '', 'enumeration' => 'aaa' }] }
 
       it 'selects the target' do
         expect(latest).to include(target)
@@ -125,5 +98,33 @@ RSpec.describe Folio::Holdings do
         expect(latest).to include(target)
       end
     end
+
+    context 'when enumeration has multiple integers' do
+      let(:target) { { 'chronology' => '', 'enumeration' => 'bd.54 t.2' } }
+      let(:options) { [{ 'chronology' => '', 'enumeration' => 'bd.54 t.1' }, target, { 'chronology' => '', 'enumeration' => 'bd.53 t.1' }] }
+
+      it 'selects the target' do
+        expect(latest).to include(target)
+      end
+    end
+
+    context 'when enumeration is text' do
+      let(:target) { { 'chronology' => '', 'enumeration' => 'ccc' } }
+      let(:options) { [{ 'chronology' => '', 'enumeration' => 'bbb' }, target, { 'chronology' => '', 'enumeration' => 'aaa' }] }
+
+      it 'selects the target' do
+        expect(latest).to include(target)
+      end
+    end
+
+    context 'when enumeration is mixed' do
+      let(:target) { { 'chronology' => '', 'enumeration' => '52' } }
+      let(:options) { [{ 'chronology' => '', 'enumeration' => 'bbb' }, target, { 'chronology' => '', 'enumeration' => 'aaa' }] }
+
+      it 'selects the target' do
+        expect(latest).to include(target)
+      end
+    end
+
   end
 end
